requirements.txt
streamlit>=1.38
python-dotenv>=1.0.1
azure-ai-documentintelligence>=1.0.2
azure-ai-vision-imageanalysis>=1.0.0
pdf2image>=1.17.0
Pillow>=10.4.0
opencv-python-headless>=4.10.0.84
numpy>=2.0.0
pandas>=2.2.2
rapidfuzz>=3.9.6
pydantic>=2.7.0
typing_extensions>=4.12.2
openai>=1.40.0
pymupdf>=1.24.4
spacy>=3.7.0
#Also need to import English model: python -m spacy download en_core_web_sm
#Also need to import models relevant to the south african contect: python -m spacy download xx_ent_wiki_sm
